Change Host header port rules slightly
RFC2616 states that Host headers should include the port, but:

A "host" without any trailing port information implies the default port
for the service requested (e.g., "80" for an HTTP URL).

Previously ssrf_filter would only conceal the port if it was 80, but it
should also do this for port 443 if the service is HTTPS. This was
leading to infinite redirects in some cases where web servers with HTTPS
would not accept <server name>:443 as a valid hostname, leading to a
redirect.