Add explicit proxy option
By default `Net::HTTP` will use `ENV['HTTP_PROXY']` for configuring
a HTTP proxy for request. However we are facing a situation where we
want to configure a proxy for a request without globally setting this
environment variable.

This change allows an explicit proxy string to be included in an
`SsrfFilter` request. It will fallback to the `ENV['HTTP_PROXY']` value
if not provided.